Added query and action layer with builders. Added analysis layer for constants.

+++ core/src/main/kotlin/maths/core/dsl/ExprDsl.kt
@@ -2,25 +2,25 @@ package maths.core.dsl
 
 import maths.core.ast.*
 
-operator fun Number.plus(e: Expr) = Add(Const(this.toDouble()), e)
-operator fun Number.minus(e: Expr) = Sub(Const(this.toDouble()), e)
-operator fun Number.times(e: Expr) = Mul(Const(this.toDouble()), e)
-operator fun Number.div(e: Expr) = Div(Const(this.toDouble()), e)
-operator fun Number.unaryMinus() = Neg(Const(this.toDouble()))
+operator fun Number.plus(e: Expr) = Add(Const(this.toInt()), e)
+operator fun Number.minus(e: Expr) = Sub(Const(this.toInt()), e)
+operator fun Number.times(e: Expr) = Mul(Const(this.toInt()), e)
+operator fun Number.div(e: Expr) = Div(Const(this.toInt()), e)
+operator fun Number.unaryMinus() = Neg(Const(this.toInt()))
 operator fun Expr.plus(other: Expr) = Add(this, other)
 operator fun Expr.minus(other: Expr) = Sub(this, other)
 operator fun Expr.times(other: Expr) = Mul(this, other)
 operator fun Expr.div(other: Expr) = Div(this, other)
 operator fun Expr.unaryMinus() = Neg(this)
-operator fun Expr.plus(other: Number) = Add(this, Const(other.toDouble()))
-operator fun Expr.minus(other: Number) = Sub(this, Const(other.toDouble()))
-operator fun Expr.times(other: Number) = Mul(this, Const(other.toDouble()))
-operator fun Expr.div(other: Number) = Div(this, Const(other.toDouble()))
+operator fun Expr.plus(other: Number) = Add(this, Const(other.toInt()))
+operator fun Expr.minus(other: Number) = Sub(this, Const(other.toInt()))
+operator fun Expr.times(other: Number) = Mul(this, Const(other.toInt()))
+operator fun Expr.div(other: Number) = Div(this, Const(other.toInt()))
 fun Expr.pow(e: Expr) = Pow(this, e)
-fun Expr.pow(n: Number) = Pow(this, Const(n.toDouble()))
-val Expr.squared get() = Pow(this, Const(2.toDouble()))
+fun Expr.pow(n: Number) = Pow(this, Const(n.toInt()))
+val Expr.squared get() = Pow(this, Const(2))
 
-val Number.c get() = Const(this.toDouble())
+val Number.c get() = Const(this.toInt())
 
 val String.v get() = Var(this)
